Rename org.eclipse.hawkbit.api -> org.eclipse.hawkbit.artifact.repository.urlhandler
_release_notes_
