+++ hawkbit-autoconfigure/src/main/java/org/eclipse/hawkbit/autoconfigure/cache/CacheAutoConfiguration.java
@@ -24,20 +24,17 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 
 /**
- * A configuration for configuring the spring {@link CacheManager} for specific
- * multi-tenancy caching. The caches between tenants must not interfere each
- * other.
+ * A configuration for configuring the spring {@link CacheManager} for specific multi-tenancy caching. The caches between
+ * tenants must not interfere each other.
  *
- * This is done by providing a special {@link TenantCacheResolver} which
- * generates a cache name included the current tenant.
+ * This is done by providing a special {@link TenancyCacheManager} which generates a cache name included the current tenant.
  */
 @Configuration
 @EnableCaching
 public class CacheAutoConfiguration {
 
     /**
-     * @return the default cache manager bean if none other cache manager is
-     *         existing.
+     * @return the default cache manager bean if none other cache manager is existing.
      */
     @Bean
     @ConditionalOnMissingBean
