+++ hawkbit-core/src/main/java/org/eclipse/hawkbit/artifact/repository/urlhandler/ApiType.java
@@ -7,7 +7,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.eclipse.hawkbit.api;
+package org.eclipse.hawkbit.artifact.repository.urlhandler;
 
 /**
  * hawkBit API type.
