+++ hawkbit-dmf/hawkbit-dmf-amqp/src/main/java/org/eclipse/hawkbit/amqp/AmqpMessageDispatcherService.java
@@ -28,11 +28,11 @@ import java.util.stream.StreamSupport;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.IterableUtils;
 import org.apache.commons.collections4.ListUtils;
-import org.eclipse.hawkbit.api.ApiType;
-import org.eclipse.hawkbit.api.ArtifactUrl;
-import org.eclipse.hawkbit.api.ArtifactUrlHandler;
-import org.eclipse.hawkbit.api.URLPlaceholder;
-import org.eclipse.hawkbit.api.URLPlaceholder.SoftwareData;
+import org.eclipse.hawkbit.artifact.repository.urlhandler.ApiType;
+import org.eclipse.hawkbit.artifact.repository.urlhandler.ArtifactUrl;
+import org.eclipse.hawkbit.artifact.repository.urlhandler.ArtifactUrlHandler;
+import org.eclipse.hawkbit.artifact.repository.urlhandler.URLPlaceholder;
+import org.eclipse.hawkbit.artifact.repository.urlhandler.URLPlaceholder.SoftwareData;
 import org.eclipse.hawkbit.dmf.amqp.api.EventTopic;
 import org.eclipse.hawkbit.dmf.amqp.api.MessageHeaderKey;
 import org.eclipse.hawkbit.dmf.amqp.api.MessageType;
